% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{auth}
\alias{auth}
\title{Authenticate License}
\usage{
auth(a1_username)
}
\arguments{
\item{a1_username}{A string containing AltairOne username.}

\item{a1_password}{A string containing AltairOne password.}
}
\value{
The output from the executable.
}
\description{
Authenticate License
}
