% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{compile}
\alias{compile}
\title{Compile Raw Code}
\usage{
compile(code, finalDSPath = "./sample_data.sas7bdat")
}
\arguments{
\item{code}{A string containing the raw code to compile.}

\item{finalDSPath}{The path to the final dataset in your sas code, where you send the .sas7bdat file}
}
\value{
The output from the executable and the log file.
}
\description{
Compile Raw Code
}
