% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{runFile}
\alias{runFile}
\title{Run a Local File}
\usage{
runFile(file_path, finalDSPath)
}
\arguments{
\item{file_path}{A string containing the path to the file to be run.}

\item{finalDSPath}{The path to the final dataset in your sas code, where you send the .sas7bdat file}
}
\value{
The output from the executable.
}
\description{
Run a Local File
}
